correccion product - cart

+++ tercera-pre-entrega/src/controllers/product.controller.js
@@ -2,27 +2,44 @@ import ProductService from "../services/product.service.js";
 import { Exception } from "../utils.js";
 
 export default class ProductController {
-    static findAll(filter = {}) {
-        return ProductService.findAll(filter);
-    }
 
     static async create(data) {
-        return ProductService.create(data);
-    }
+        console.log('Creando el nuevo producto.');
+        const newProduct = await ProductService.create(data);
+        console.log('Producto creado corretamente.');
+        return newProduct;
+      }
     
-    static async findById(pid) {
+      static async get(query = {}) {
+        const product = await ProductService.findAll(query); 
+        return product;
+      }
+    
+      static async getById(pid) {
         const product = await ProductService.findById(pid);
         if (!product) {
             throw new Exception('No existe el producto', 404);
         }
         return product;
-    }
+      }
     
-    static updateById(pid, data) {
-        return ProductService.updateById(pid, data);
-     }
+      static async updateById(pid, data) {
+        await ProductController.getById(pid);
+        if(!product) {
+          throw new Exception('No existe el producto.', 404);
+        };
+        console.log('Actualizando el producto.');
+        await ProductService.updateById(pid, data);
+        console.log('Actualizado el producto corretamente.');
+      }
     
-    static deleteById(pid) {
-        return ProductService.deleteById(pid);
-    }
+      static async deleteById(pid) {
+        await ProductController.getById(pid);
+        if(!product) {
+          throw new Exception('No existe el producto', 404);
+        };
+        console.log('Eliminando el producto.');
+        await ProductService.deleteById(pid);
+        console.log('Producto eliminado corretamente.');
+      }
 }
